requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
behave==1.2.6
certifi==2024.2.2
charset-normalizer==3.3.2
Faker==24.7.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pluggy==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0